avl
currently there is a slight problem with the code. the calculated lift of the wing is way too low and doesn't equal the total lift required. In varying the angle of alpha, it is possible to obtain values of total lift for the wing that seem more coherent. will explore tonight/ Wednesday :sweat_smile: